Feature/default regenerators (#31)
* Update SaveMapController.java

Set default regenerators to 100 when creating new nodes

* Update SaveMapController.java

Set default for Number of Regenerators to 100 without filling in the text field. Clarified dialog boxes.

* Final Commit for Default Regenerators

Fixed prompt text for number of regenerators

* Final Commit

The actual final commit for this branch. Fixed all dialog boxes and set default for regenerators

* Refactoring, changed some placeholder texts, extracted the hard coded string to the bundle files and increased the width of the number of regenerators input

* Small GUI Fix for width on Regenerator Input

set width to 230 from 200 as on some computers, it cuts off some prompt text if set to 200